Remove "The Song Remains The Same" from clean_midi
In the Million Song Dataset, all of the songs in the Led Zeppelin album
"The Song Remains The Same" (2007 reissue) are called something like
"Song Name  (2007 Remastered Live Version From The Song Remains The Same)".
When doing text-based matching, when two MIDI files share a common audio
file match, their collection of matches are merged.  This is for cases
like when "Under Pressure" appears in the clean MIDI subset as both a
Queen and a David Bowie song - it ensures we are truly splitting by
song.  However, it also might result in some spurious matches.  In the
case of "The Song Remains the Same", it's particularly bad - every
version of every Led Zeppelin song which appears on the 2007 remaster of
"The Song Remains The Same" gets matched to every other one, resulting
in 700 matches, most of which are spurious.  There are a few ways to
handle this.  One would be some sort of ad-hoc filter for this king of
thing, which is ugly.  Another would be to delete the matches from the
text_matches.js file which results from doing the text matching, which
is not really reproducible.  A third would be to simply delete all MIDI
files of "The Song Remains The Same" from the clean MIDI subset.  I
chose this last option because it turns out none of them align
particularly well to the single actual recording of "The Song Remains
The Same" in the Million Song Dataset.  So, removing them doesn't
actually sacrifice us any quality training data, and it also maintains
reproducibility and doesn't require any unseemly hacks.  This commit
reflects the removal.