Make m_XY and alpha_XY param ranges start non-zero
When they are zero, failure is guaranteed because there is no "negative
example" cost.  The network is able to decrease the cost by simply
mapping everything to the exact same hash code, and further saturating
the tanh units to get incremental decreases in cost, which is useless.
So, if we bump up these ranges to non-zero, this hopefully will never
happen, and we won't miss out on any good parts of the parameter search
space.