Add script for precomputing hash sequences
We are only precomputing hash sequences for MIDI files in the
development and test splits, and for the entire MSD.  It turns out that
reading a single hash-sequence h5 file with hickle takes about 3
milliseconds, so we're using messagepack instead, which is a broadly
compatible format which only takes about 10 microseconds to read.
Sad to break the consistency of h5 files, but oh well.