Use np.nanmin for finding smallest objective value
The objective value for trials can be NaN, which to Spearmint denotes
"this is a very bad region of the parameter space".  However, using
np.min on an array (or list) of values with a NaN will return NaN, not
the smallest non-NaN value.  You need to use np.nanmin for that.