Increase negative_importance range in dhs
This reflects the changes in
https://github.com/craffel/dhs/commit/4664fbcbfd0e3f98b67c9c4c925766d4f5ebe64f
https://github.com/craffel/dhs/commit/db94b4e99759ecee9da1bfa259ca4a44d310639f
These two changes mean that the cost terms other than negative
importance are now n_bits times as high as they were.  So, we should
increase the range to [0, n_bits] instead of [0, 1].  But, in practice,
I didn't see any settings for this hyperparameter over .5, so setting it
to 20 seems fine.