Crop synthesized audio in fast_fluidsynth
Sometimes (maybe .1% of the time), fluidsynth would pad a bunch of
silence to the end of a synthesis.  This resulted in a CQT with a big
silent block, which would result in false matches to the 2 (+?) tracks
in the MSD whose 7digital preview clips were silent (they have hidden
tracks).  Probably causes issues elsewhere, too.