Add DHS piano roll experiment
This experiment is basically the same as DHS, except that instead of
using a CQT of a synthesized MIDI, it uses a MIDI piano roll matrix.
This tests whether the model can learn a shared representation of
multimodal data.  Since the feature representation is different, we need
to generate an additional training/validation set for piano rolls, and
then when precomputing hash sequences we need to generate piano rolls
from the dev/test sets on the fly before hashing them.